requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.38.0
networkx==3.5
numpy==1.26.4
openai==1.78.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.45.0
streamlit-option-menu==0.4.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
zstandard==0.23.0
